Review 1leg part 1/4 - Power schematics
This commit contains the review of the power schematics as of 31/08/2020.
Comments and questions are done via blue containers directly on the schematics.
Some minor net labels movements have been made for better readeability.
Bi-directional TVS were added to protect all current measurement units. These are comprised of components already available in KiCad's library and should be revisisted.

Out of the sub-functions, here's the review:
    Input filter - film capacitor missing
    Sensors - All ok
    Inverter leg - Switches datasheet link broken, no power diode reference or datasheet, no inductor datasheet reference
    Connectors (input/output) - No datasheet reference. Footprint not verified.
    N/GND connection - All ok
    Multi-board component - All connections ok